Basic Enemy Player Detection Added New Enemy Controller for shoot enemies Reduced background brightness added light source Changed background scrolling speed - horizontal and vertical Added all animator controllers to enemies basic animations for shooting enemies Fixed Trigger Event Bug that allowed lasers to open doors

+++ Assets/Standard Assets/Scripts/TriggerEvent.cs
@@ -10,7 +10,7 @@ public class TriggerEvent : MonoBehaviour {
 			Destroy (other.gameObject); // change this for pushable scripting
 
 		if (gameObject.tag == "Player Object")
-			if (requiresItemID != null)
+			if (requiresItemID != null && other.tag == "Player")
 				if (ItemDatabase.HasItem(requiresItemID)){
 					if (removeItemAfterUse)
 						ItemDatabase.RemoveItem(ItemDatabase.GetItem(requiresItemID));
